\encoding{UTF-8}
\name{GSE45827expr}
\alias{GSE45827expr}
\docType{data}
\title{Gene expression profiles of the data set GSE45827}
\description{
  A gene expression profiles contains breast cancer PAM50 subtypes samples and normal samples.
}
\usage{data("GSE45827expr")}
\details{
  This data sets were downloaded from the GEO database (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE45827), which include 130 breast cancers (41 Basal, 30 HER2, 29 Luminal A and 30 Luminal B) as well as 11 normal tissues samples. The probes in the gene expression profiles were converted into gene symbol. For repeated probes or multiple probes corresponding to a gene symbol, we aggregated their expression values by median.
}
\keyword{datasets}
